+++ server/routes/location.route.js
@@ -1,9 +1,9 @@
 import express from 'express';
 import Location from '../models/location.model.js';
-import Review from '../models/review.model.js';
-import jwt from 'jsonwebtoken';
 import dotenv from 'dotenv';
+
 dotenv.config();
+
 const router = express.Router();
 
 //Search Route
@@ -197,114 +197,6 @@ router.get('/filter', async (req, res) => {
     } catch (error) {
         console.error('Error fetching locations:', error);
         res.status(500).json({ message: 'Error fetching locations', error });
-    }
-});
-
-
-const authenticateUser = (req, res, next) => {
-    const token = req.header('Authorization')?.replace('Bearer ', '');
-    
-    if (!token) {
-        return res.status(401).json({ success: false, message: 'User not authenticated' });
-    }
-
-    try {
-        const decoded = jwt.verify(token, process.env.SECRET_ACCESS_TOKEN);  
-        req.user = decoded;
-        next();
-    } catch (error) {
-        res.status(401).json({ success: false, message: 'Invalid or expired token' });
-    }
-};
-
-
-router.post('/review/:locationId', authenticateUser, async (req, res) => {
-    const { locationId } = req.params;
-    const { review } = req.body;
-    const userId = req.user ? req.user.id : null;  
-
-    if (!userId) {
-        return res.status(401).json({ success: false, message: 'User not authenticated' });
-    }
-
-    try {
-        const newReview = await Review.create({ locationId, userId, review });
-        res.status(201).json({ success: true, review: newReview });
-    } catch (error) {
-        res.status(500).json({ success: false, message: error.message });
-    }
-});
-
-
-router.get('/review/:locationId', async (req, res) => {
-    const { locationId } = req.params;
-
-    try {
-        const reviews = await Review.find({ locationId });
-
-        if (!reviews || reviews.length === 0) {
-            return res.status(404).json({ success: false, message: 'No reviews found for this location' });
-        }
-
-        res.status(200).json({ success: true, reviews });
-    } catch (error) {
-        res.status(500).json({ success: false, message: error.message });
-    }
-});
-
-
-router.delete('/review/:locationId/:reviewId', authenticateUser, async (req, res) => {
-    const { locationId, reviewId } = req.params;
-    const userId = req.user ? req.user.id : null;  
-
-    if (!userId) {
-        return res.status(401).json({ success: false, message: 'User not authenticated' });
-    }
-
-    try {
-        
-        const review = await Review.findByIdAndDelete(reviewId, {
-            conditions: { userId, locationId }
-        });
-
-        if (!review) {
-            return res.status(404).json({ success: false, message: 'Review not found or not authorized to delete this review' });
-        }
-
-        res.status(200).json({ success: true, message: 'Review deleted successfully' });
-    } catch (error) {
-        res.status(500).json({ success: false, message: error.message });
-    }
-});
-
-
-router.put('/review/:locationId/:reviewId', authenticateUser, async (req, res) => {
-    const { locationId, reviewId } = req.params;
-    const { review } = req.body;
-    const userId = req.user ? req.user.id : null; 
-
-    if (!userId) {
-        return res.status(401).json({ success: false, message: 'User not authenticated' });
-    }
-
-    try {
-        const existingReview = await Review.findOne({ 
-            _id: reviewId,
-            locationId: locationId, 
-            userId: userId
-        });
-
-        if (!existingReview) {
-            return res.status(404).json({ success: false, message: 'Review not found' });
-        }
-
-        existingReview.review = review || existingReview.review;
-
-        await existingReview.save();
-
-        res.status(200).json({ success: true, review: existingReview });
-    } catch (error) {
-        res.status(500).json({ success: false, message: error.message });
     }
 });
